refactor: 내 프로필, 상대 프로필의 중복 화면을 별도 폴더로 분리하기
