+++ components/common/profile/index.ts
@@ -1,0 +1,4 @@
+export { default as Content } from "./content";
+export { default as MenuTabs, Tab as MenuTabType } from "./menu-tabs";
+export { default as ProfileImageBox } from "./profile-image-box";
+export { default as UserActivitySummary } from "./user-activity-summary";
