+++ components/my-profile/user/index.ts
@@ -1,5 +1,3 @@
 export { default as ControlButtons } from "./control-buttons";
 export { default as Header } from "./header";
-export { default as MenuTabs, Tab as MenuTabType } from "./menu-tabs";
-export { default as Content } from "./content";
 export { default as ProfileSelectSheet } from "./profile-select-sheet";
