+++ components/other-profile/user/index.ts
@@ -1,4 +1,3 @@
 export { default as ControlButtons } from "./control-buttons";
 export { default as Header } from "./header";
-export { default as MenuTabs, Tab as MenuTabType } from "./menu-tabs";
 export { default as ProfileDashboardSheet } from "./profile-dashboard-sheet";
